+++ Sentry.podspec
@@ -41,6 +41,6 @@ Pod::Spec.new do |s|
         "Sources/SentryCrash/**/*.{h,hpp,m,mm,c,cpp}", "Sources/Swift/Sentry.swift"
         
       sp.public_header_files =
-        "Sources/Sentry/Public/*.h", "Sources/Sentry/include/PrivateSentrySDKOnly.h"
+        "Sources/Sentry/Public/*.h", "Sources/Sentry/include/HybridPublic/*.h"
   end
 end
