+++ Sources/Sentry/include/HybridPublic/SentryAppStartMeasurement.h
