+++ Sources/Sentry/include/HybridPublic/SentryEnvelope.h
