+++ Sources/Sentry/include/HybridPublic/SentryEnvelopeItemType.h
