+++ Sources/Sentry/include/HybridPublic/SentryScreenFrames.h
